chore: Break network notifications into modules
